change algotrading dynamics
bugfixes for skip connections

improve dtype specs

+++ src/rlstack/views.py
@@ -74,7 +74,7 @@ def pad_last_sequence(x: torch.Tensor, size: int, /) -> TensorDict:
     pad = size - T
     if pad > 0:
         F = x.shape[2:]
-        padding = torch.zeros(B, pad, *F, device=x.device)
+        padding = torch.zeros(B, pad, *F, device=x.device, dtype=x.dtype)
         x = torch.cat([padding, x], 1)
         padding_mask = torch.zeros(B, size, device=x.device, dtype=torch.bool)
         padding_mask[:, :pad] = True
@@ -108,7 +108,7 @@ def pad_whole_sequence(x: torch.Tensor, size: int, /) -> TensorDict:
     B, T = x.shape[:2]
     F = x.shape[2:]
     pad = RollingWindow.drop_size(size)
-    padding = torch.zeros(B, pad, *F, device=x.device)
+    padding = torch.zeros(B, pad, *F, device=x.device, dtype=x.dtype)
     x = torch.cat([padding, x], 1)
     padding_mask = torch.zeros(B, T + pad, device=x.device, dtype=torch.bool)
     padding_mask[:, :pad] = True
